refactor: change DbFlavour enum to be public.
