+++ core/src/integrationTest/kotlin/net/samyn/kapper/DbTypeConverter.kt
@@ -1,6 +1,5 @@
 package net.samyn.kapper
 
-import net.samyn.kapper.internal.DbFlavour
 import java.util.UUID
 
 private val specialTypes =
