+++ core/src/main/kotlin/net/samyn/kapper/DbFlavour.kt
@@ -1,0 +1,10 @@
+package net.samyn.kapper
+
+enum class DbFlavour {
+    POSTGRESQL,
+    MYSQL,
+    SQLITE,
+    ORACLE,
+    MSSQLSERVER,
+    UNKNOWN,
+}
