+++ core/src/test/kotlin/net/samyn/kapper/internal/KapperImplQueryTest.kt
@@ -7,6 +7,7 @@ import io.mockk.mockk
 import io.mockk.mockkStatic
 import io.mockk.unmockkAll
 import io.mockk.verify
+import net.samyn.kapper.DbFlavour
 import net.samyn.kapper.Field
 import net.samyn.kapper.KapperQueryException
 import net.samyn.kapper.KapperResultException
