+++ core/src/test/kotlin/net/samyn/kapper/internal/QueryExecutionTest.kt
@@ -8,6 +8,7 @@ import io.mockk.mockkStatic
 import io.mockk.unmockkAll
 import io.mockk.verify
 import net.samyn.kapper.Args
+import net.samyn.kapper.DbFlavour
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.parallel.Isolated
