+++ core/src/test/kotlin/net/samyn/kapper/internal/automapper/FieldsConverterTest.kt
@@ -3,6 +3,7 @@ package net.samyn.kapper.internal.automapper
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
+import net.samyn.kapper.DbFlavour
 import net.samyn.kapper.Field
 import org.junit.jupiter.api.Test
 import java.sql.JDBCType
@@ -17,8 +18,8 @@ class FieldsConverterTest {
     fun `should convert all fields to column values`() {
         val fields =
             mapOf(
-                "id" to Field(1, JDBCType.INTEGER, "Int", net.samyn.kapper.internal.DbFlavour.UNKNOWN),
-                "name" to Field(2, JDBCType.VARCHAR, "String", net.samyn.kapper.internal.DbFlavour.UNKNOWN),
+                "id" to Field(1, JDBCType.INTEGER, "Int", DbFlavour.UNKNOWN),
+                "name" to Field(2, JDBCType.VARCHAR, "String", DbFlavour.UNKNOWN),
             )
         every { sqlTypesConverter.convert(fields["id"]!!, resultSet) } returns 42
         every { sqlTypesConverter.convert(fields["name"]!!, resultSet) } returns "Bruce"
@@ -41,7 +42,7 @@ class FieldsConverterTest {
     fun `should allow null values from converter`() {
         val fields =
             mapOf(
-                "email" to Field(1, JDBCType.VARCHAR, "String", net.samyn.kapper.internal.DbFlavour.UNKNOWN),
+                "email" to Field(1, JDBCType.VARCHAR, "String", DbFlavour.UNKNOWN),
             )
         every { sqlTypesConverter.convert(fields["email"]!!, resultSet) } returns null
         val result = fieldsConverter.convert(resultSet, fields)
