+++ core/src/test/kotlin/net/samyn/kapper/internal/automapper/KotlinDataClassMapperTest.kt
@@ -6,10 +6,10 @@ import io.kotest.matchers.string.shouldContain
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
+import net.samyn.kapper.DbFlavour
 import net.samyn.kapper.Field
 import net.samyn.kapper.KapperMappingException
 import net.samyn.kapper.internal.AutoConverter
-import net.samyn.kapper.internal.DbFlavour
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
